refactor: Move stabs and labels to course-header
